Rename file and remove `__init__` magic
