--- readthedocs/notifications/__init__.py
@@ -1,6 +1,0 @@
-"""Notification system to send email/site messages to users."""
-from .notification import EmailNotification
-
-__all__ = (
-    "EmailNotification",
-)
